Modify Predictors corresponding files
