+++ configs/test/acdc_sisr_config.yaml
@@ -6,24 +6,20 @@ dataset:
     name: 'AcdcSISRDataset'
     kwargs:
         data_dir: 'path/to/data'
+        downscale_factor: value
         transforms:
-            - name: 'RandomElasticDeformation'
-              kwargs:
-                  prob: 0.75
-            - name: 'RandomCrop'
-              kwargs:
-                  size: [128, 128]
-        post_transforms:
             - name: 'Normalize'
               kwargs:
-                  means: [53.434]
-                  stds: [47.652]
+                  means: [54.089]
+                  stds: [48.084]
             - name: 'ToTensor'
-        degrade:
-            - name: 'Degrade'
+        augments:
+            - name: 'RandomHorizontalFlip'
+            - name: 'RandomVerticalFlip'
+            - name: 'RandomCropPatch'
               kwargs:
-                  method: 'FFT'
-                  downscale_factor: 4
+                  size: [32, 32]
+                  ratio: value
 
 dataloader:
     name: 'Dataloader'
@@ -52,4 +48,4 @@ predictor:
     kwargs:
         device: 'cuda:0'
         saved_dir: 'path/to/predicted/results'
-        export_prediction: true
+        exported: true
