+++ src/runner/predictors/__init__.py
@@ -2,5 +2,4 @@ from .acdc_sisr_predictor import AcdcSISRPredictor
 from .acdc_sisr_bicubic_predictor import AcdcSISRBicubicPredictor
 from .acdc_sisr_srfb_predictor import AcdcSISRSRFBPredictor
 from .acdc_misr_predictor import AcdcMISRPredictor
-from .acdc_misr_duf_predictor import AcdcMISRDUFPredictor
 from .acdc_vsr_predictor import AcdcVSRPredictor
