+++ src/runner/predictors/acdc_sisr_predictor.py
@@ -11,20 +11,19 @@ from src.runner.predictors.base_predictor import BasePredictor
 
 
 class AcdcSISRPredictor(BasePredictor):
-    """The ACDC predictor for the Single-Image Super Resolution.
+    """The ACDC predictor for the Single-Image Super-Resolution.
     Args:
         saved_dir (str): The directory to save the predicted videos, images and metrics (default: None).
-        export_prediction (bool): Whether to export the predicted video, images and metrics (default: False).
+        exported (bool): Whether to export the predicted video, images and metrics (default: False).
     """
-    def __init__(self, saved_dir=None, export_prediction=False, **kwargs):
+    def __init__(self, saved_dir=None, exported=False, **kwargs):
         super().__init__(**kwargs)
-        if export_prediction:
-            if self.test_dataloader.batch_size != 1:
-                raise ValueError(f'The batch size should be 1 if export_prediction is True. Got {self.test_dataloader.batch_size}.')
-            #if self.test_dataloader.shuffle is True:
-            #    raise ValueError('The shuffle should be False if export_prediction is True.')
+        if self.test_dataloader.batch_size != 1:
+            raise ValueError(f'The testing batch size should be 1. Got {self.test_dataloader.batch_size}.')
+
+        if exported:
             self.saved_dir = Path(saved_dir)
-        self.export_prediction = export_prediction
+        self.exported = exported
 
     def predict(self):
         """The testing process.
@@ -34,7 +33,7 @@ class AcdcSISRPredictor(BasePredictor):
                       total=len(self.test_dataloader),
                       desc='testing')
 
-        if self.export_prediction:
+        if self.exported:
             videos_dir = self.saved_dir / 'videos'
             imgs_dir = self.saved_dir / 'imgs'
             csv_path = self.saved_dir / 'results.csv'
@@ -57,9 +56,9 @@ class AcdcSISRPredictor(BasePredictor):
                 loss = (torch.stack(losses) * self.loss_weights).sum()
                 metrics = self._compute_metrics(output, target)
 
-                if self.export_prediction:
-                    path = self.test_dataloader.dataset.data_paths[index]
-                    filename = path.parts[-1].split('.')[0]
+                if self.exported:
+                    lr_path, hr_path = self.test_dataloader.dataset.data[index]
+                    filename = lr_path.parts[-1].split('.')[0]
                     patient, _, sid, fid = filename.split('_')
 
                     _losses = [loss.item() for loss in losses]
@@ -92,7 +91,7 @@ class AcdcSISRPredictor(BasePredictor):
             trange.set_postfix(**dict((key, f'{value / count: .3f}') for key, value in log.items()))
 
         # Save the results.
-        if self.export_prediction:
+        if self.exported:
             with open(csv_path, 'w', newline='') as csvfile:
                 writer = csv.writer(csvfile)
                 writer.writerows(results)
@@ -134,9 +133,7 @@ class AcdcSISRPredictor(BasePredictor):
         Returns:
             metrics (list of torch.Tensor): The computed metrics.
         """
-        # Do the denormalization to [0-255] before computing the metric.
         output, target = self._denormalize(output), self._denormalize(target)
-
         metrics = [metric_fn(output, target) for metric_fn in self.metric_fns]
         return metrics
 
@@ -151,8 +148,8 @@ class AcdcSISRPredictor(BasePredictor):
                 writer.append_data(img)
 
     @staticmethod
-    def _denormalize(imgs, mean=53.434, std=47.652):
-        """Denormalize the images to [0-255].
+    def _denormalize(imgs, mean=54.089, std=48.084):
+        """Denormalize the images.
         Args:
             imgs (torch.Tensor) (N, C, H, W): Te images to be denormalized.
             mean (float): The mean of the training data.
