+++ src/runner/predictors/acdc_vsr_predictor.py
@@ -12,16 +12,18 @@ from src.runner.predictors.base_predictor import BasePredictor
 
 class AcdcVSRPredictor(BasePredictor):
     """The ACDC predictor for the Video Super-Resolution.
+    Args:
+        saved_dir (str): The directory to save the predicted videos, images and metrics (default: None).
+        exported (bool): Whether to export the predicted video, images and metrics (default: False).
     """
-    def __init__(self, saved_dir=None, export_prediction=False, **kwargs):
+    def __init__(self, saved_dir=None, exported=False, **kwargs):
         super().__init__(**kwargs)
-        if export_prediction:
-            if self.test_dataloader.batch_size != 1:
-                raise ValueError(f'The batch size should be 1 if export_prediction is True. Got {self.test_dataloader.batch_size}.')
-            #if self.test_dataloader.shuffle is True:
-            #    raise ValueError('The shuffle should be False if export_prediction is True.')
+        if self.test_dataloader.batch_size != 1:
+            raise ValueError(f'The testing batch size should be 1. Got {self.test_dataloader.batch_size}.')
+
+        if exported:
             self.saved_dir = Path(saved_dir)
-        self.export_prediction = export_prediction
+        self.exported = exported
 
     def predict(self):
         """The testing process.
@@ -31,7 +33,7 @@ class AcdcVSRPredictor(BasePredictor):
                       total=len(self.test_dataloader),
                       desc='testing')
 
-        if self.export_prediction:
+        if self.exported:
             videos_dir = self.saved_dir / 'videos'
             imgs_dir = self.saved_dir / 'imgs'
             csv_path = self.saved_dir / 'results.csv'
@@ -55,9 +57,9 @@ class AcdcVSRPredictor(BasePredictor):
                 loss = (losses.mean(dim=0) * self.loss_weights).sum()
                 metrics = self._compute_metrics(outputs, targets)
 
-                if self.export_prediction:
-                    path = self.test_dataloader.dataset.data_paths[index]
-                    filename = path.parts[-1].split('.')[0]
+                if self.exported:
+                    lr_path, hr_path = self.test_dataloader.dataset.data[index]
+                    filename = lr_path.parts[-1].split('.')[0]
                     patient, _, sid = filename.split('_')
 
                     filename = filename.replace('2d+1d', '2d').replace('sequence', 'slice')
@@ -90,7 +92,7 @@ class AcdcVSRPredictor(BasePredictor):
             trange.set_postfix(**dict((key, f'{value / count: .3f}') for key, value in log.items()))
 
         # Save the results.
-        if self.export_prediction:
+        if self.exported:
             with open(csv_path, 'w', newline='') as csvfile:
                 writer = csv.writer(csvfile)
                 writer.writerows(results)
@@ -135,7 +137,6 @@ class AcdcVSRPredictor(BasePredictor):
         Returns:
             metrics (list of torch.Tensor): The computed metrics.
         """
-        # Do the denormalization before computing the metric.
         outputs = list(map(self._denormalize, outputs))
         targets = list(map(self._denormalize, targets))
 
@@ -172,12 +173,13 @@ class AcdcVSRPredictor(BasePredictor):
                 writer.append_data(img)
 
     @staticmethod
-    def _denormalize(imgs, mean=53.434, std=47.652):
-        """Denormalize the images to [0-255].
+    def _denormalize(imgs, mean=54.089, std=48.084):
+        """Denormalize the images.
         Args:
             imgs (torch.Tensor) (N, C, H, W): Te images to be denormalized.
             mean (float): The mean of the training data.
             std (float): The standard deviation of the training data.
+
         Returns:
             imgs (torch.Tensor) (N, C, H, W): The denormalized images.
         """
